+++ App/AbletonLiveStopwatch/EmbeddedRemoteScript/LiveStopwatch_Clip_Watcher/__init__.py
@@ -1,0 +1,5 @@
+from .live_stopwatch_clip_watcher import LiveStopwatchClipWatcher
+
+
+def create_instance(c_instance):
+    return LiveStopwatchClipWatcher(c_instance)