+++ App/AbletonLiveStopwatch/EmbeddedRemoteScript/SIGNAL_FLOW_Clip_Watcher/__init__.py
@@ -1,5 +1,0 @@
-from .signal_flow_clip_watcher import SignalFlowClipWatcher
-
-
-def create_instance(c_instance):
-    return SignalFlowClipWatcher(c_instance)